Improve how user is passed from users table to popups; fix VHT Supervised dropdown bugs
- Fix bug: if VHT changed to CHO, that user could be selected in the "VHT Supervises" list and be supervising themselves

- Fix bug: if VHT was changed to another role, that VHT couldn't be removed from the "supervises" list of other users

+++ client/src/pages/admin/manageUsers/DeleteUser.tsx
@@ -11,21 +11,26 @@ import { EndpointEnum } from 'src/server';
 import { BASE_URL } from 'src/server/utils';
 import { Toast } from 'src/shared/components/toast';
 import { apiFetch } from 'src/shared/utils/api';
+import { IUser } from 'src/types';
 
 interface IProps {
   open: boolean;
   onClose: () => void;
-  userId: number;
-  name: string;
+  deleteUser: IUser | undefined;
 }
 
-const DeleteUser = ({ open, onClose, userId, name }: IProps) => {
+const DeleteUser = ({ open, onClose, deleteUser }: IProps) => {
   const [submitError, setSubmitError] = useState(false);
   const [submitSuccess, setSubmitSuccess] = useState(false);
+  const name = deleteUser?.firstName ?? '';
 
   const handleDelete = async () => {
+    if (!deleteUser) {
+      return;
+    }
+
     try {
-      const url = BASE_URL + EndpointEnum.USER + String(userId);
+      const url = BASE_URL + EndpointEnum.USER + String(deleteUser.userId);
       const resp = await apiFetch(url, {
         method: 'DELETE',
       });
